more style change
Finally fixed the line under the navigation, other style changes like fixing signin and stuff again. plus chaning the account page and adding some more to the home page. still have to fix alot more of the pages but it is kind of hard considering I dont really know what to have on the page itself.