trying to fix single user print
this thing is kind of stupid, because if sentence wont work inside JSX so I kinda have to jerryrig something that will work but wont be the best solution to be honest.
I@ll probbably fix something better eventually, but have done alot on this at the moment and I just really want to move on. if you actually read this then cudos to you sir, but I have to be done with it now so goodbye and see you around another time in  another text ;)